+++ releases/ChangeLog.md
@@ -1,6 +1,10 @@
 ## 5.2.3 - 2024-XX-XX
 
 - Bugfix: Saving passphrase on SD Card caused a freeze that required reboot
+- Bugfix: Properly handle and finalize framing error response
+- Bugfix: `Brick Me` option for `If Wrong` PIN caused yikes
+- Bugfix: Do not allow non-ascii or ascii non-printable characters in multisig
+  wallet name
 
 ## 5.2.2 - 2023-12-21